+++ sparkle/CLI/configure_solver.py
@@ -3,10 +3,7 @@
 from __future__ import annotations
 import argparse
 import sys
-import os
 import math
-
-from pandas import DataFrame
 
 from runrunner import Runner
 
@@ -19,7 +16,6 @@ from sparkle.CLI.help import argparse_custom as ac
 
 from sparkle.platform.settings_objects import SettingState
 from sparkle.structures import PerformanceDataFrame, FeatureDataFrame
-from sparkle.configurator import implementations as configurator_implementations
 from sparkle.solver import Solver
 from sparkle.instance import Instance_Set
 
@@ -89,6 +85,7 @@ def main(argv: list[str]) -> None:
     """Main function of the configure solver command."""
     # Log command call
     sl.log_command(sys.argv)
+    check_for_initialise()
 
     parser = parser_function()
 
@@ -118,39 +115,8 @@ def main(argv: list[str]) -> None:
     use_features = args.use_features
     run_on = gv.settings().get_run_on()
 
-    check_for_initialise()
-
     configurator = gv.settings().get_general_sparkle_configurator()
     configurator_settings = gv.settings().get_configurator_settings(configurator.name)
-    if use_features and configurator.name == configurator_implementations.SMAC2.__name__:
-        feature_data = FeatureDataFrame(gv.settings().DEFAULT_feature_data_path)
-
-        data_dict = {}
-        feature_data_df = feature_data.dataframe
-
-        for label, row in feature_data_df.iterrows():
-            # os.path.split(os.path.split(label)[0])[1] gives the dir/instance set name
-            if os.path.split(os.path.split(label)[0])[1] == instance_set_train.name:
-                if row.empty:
-                    print("No feature data exists for the given training set, please "
-                          "run add_feature_extractor.py, then compute_features.py")
-                    sys.exit(-1)
-
-                new_label = (f"../../../instances/{instance_set_train.name}/"
-                             + os.path.split(label)[1])
-                data_dict[new_label] = row
-
-        feature_data_df = DataFrame.from_dict(data_dict, orient="index",
-                                              columns=feature_data_df.columns)
-
-        if feature_data.has_missing_value():
-            print("You have unfinished feature computation jobs, please run "
-                  "`sparkle compute features`")
-            sys.exit(-1)
-
-        for index, column in enumerate(feature_data_df):
-            feature_data_df.rename(columns={column: f"Feature{index+1}"}, inplace=True)
-        configurator_settings.update({"feature_data_df": feature_data_df})
 
     sparkle_objectives =\
         gv.settings().get_general_sparkle_objectives()
@@ -164,6 +130,23 @@ def main(argv: list[str]) -> None:
                   "Adding to data frame.")
             performance_data.add_objective(objective.name)
 
+    if use_features:
+        feature_data = FeatureDataFrame(gv.settings().DEFAULT_feature_data_path)
+        # Check that the train instance set is in the feature data frame
+        invalid = False
+        remaining_instance_jobs =\
+            set([instance for instance, _, _ in feature_data.remaining_jobs()])
+        for instance in instance_set_train.instance_paths:
+            if str(instance) not in feature_data.instances:
+                print(f"ERROR: Train Instance {instance} not found in feature data.")
+                invalid = True
+            elif instance in remaining_instance_jobs:  # Check jobs
+                print(f"ERROR: Features have not been computed for instance {instance}.")
+                invalid = True
+        if invalid:
+            sys.exit(-1)
+        configurator_settings.update({"feature_data": feature_data})
+
     sbatch_options = gv.settings().get_slurm_extra_options(as_args=True)
     config_scenario = configurator.scenario_class()(
         solver, instance_set_train, sparkle_objectives,
@@ -172,15 +155,16 @@ def main(argv: list[str]) -> None:
     # Run the default configuration
     remaining_jobs = performance_data.get_job_list()
     relevant_jobs = []
-    for job in remaining_jobs:
-        if job[-1] != str(solver.directory):
+    for instance, run_id, solver_id in remaining_jobs:
+        # NOTE: This run_id skip will not work if we do multiple runs per configuration
+        if run_id != 1 or solver_id != str(solver.directory):
             continue
         configuration = performance_data.get_value(
-            job[2], job[0], sparkle_objectives[0].name, run=job[1],
+            solver_id, instance, sparkle_objectives[0].name, run=run_id,
             solver_fields=[PerformanceDataFrame.column_configuration])
         # Only run jobs with the default configuration
         if not isinstance(configuration, str) and math.isnan(configuration):
-            relevant_jobs.append(job)
+            relevant_jobs.append((instance, run_id, solver_id))
 
     # Expand the performance dataframe so it can store the configuration
     performance_data.add_runs(configurator_runs,
@@ -213,6 +197,8 @@ def main(argv: list[str]) -> None:
             cutoff_time=config_scenario.cutoff_time,
             log_dir=config_scenario.validation,
             base_dir=sl.caller_log_dir,
+            job_name=f"Default Configuration: {solver.name} Validation on "
+                     f"{instance_set_train.name}",
             run_on=run_on)
         dependency_job_list.append(default_job)
 
@@ -226,24 +212,28 @@ def main(argv: list[str]) -> None:
         gv.latest_scenario().set_config_instance_set_test(instance_set_test.directory)
         # Schedule test set jobs
         if args.test_set_run_all_configurations:
+            # TODO: Schedule test set runs for all configurations
+            print("Running all configurations on test set is not implemented yet.")
             pass
         else:
             # We place the results in the index we just added
             run_index = list(set([performance_data.get_instance_num_runs(str(i))
                                   for i in instance_set_test.instance_paths]))
-        test_set_job = solver.run_performance_dataframe(
-            instance_set_test,
-            run_index,
-            performance_data,
-            cutoff_time=config_scenario.cutoff_time,
-            objective=config_scenario.sparkle_objective,
-            train_set=instance_set_train,
-            sbatch_options=sbatch_options,
-            log_dir=config_scenario.validation,
-            base_dir=sl.caller_log_dir,
-            dependencies=dependency_job_list,
-            run_on=run_on)
-        dependency_job_list.append(test_set_job)
+            test_set_job = solver.run_performance_dataframe(
+                instance_set_test,
+                run_index,
+                performance_data,
+                cutoff_time=config_scenario.cutoff_time,
+                objective=config_scenario.sparkle_objective,
+                train_set=instance_set_train,
+                sbatch_options=sbatch_options,
+                log_dir=config_scenario.validation,
+                base_dir=sl.caller_log_dir,
+                dependencies=dependency_job_list,
+                job_name=f"Best Configuration: {solver.name} Validation on "
+                         f"{instance_set_test.name}",
+                run_on=run_on)
+            dependency_job_list.append(test_set_job)
     else:
         # Set to default to overwrite possible old path
         gv.latest_scenario().set_config_instance_set_test()