+++ sparkle/configurator/implementations/irace.py
@@ -6,7 +6,7 @@ from pathlib import Path
 
 from sparkle.configurator.configurator import Configurator, ConfigurationScenario
 from sparkle.solver import Solver
-from sparkle.structures import PerformanceDataFrame
+from sparkle.structures import PerformanceDataFrame, FeatureDataFrame
 from sparkle.instance import InstanceSet, Instance_Set
 from sparkle.types import SparkleObjective, resolve_objective
 
@@ -205,6 +205,7 @@ class IRACEScenario(ConfigurationScenario):
                  first_test: int = None,
                  mu: int = None,
                  max_iterations: int = None,
+                 feature_data: FeatureDataFrame = None,
                  )\
             -> None:
         """Initialize scenario paths and names.
@@ -237,6 +238,8 @@ class IRACEScenario(ConfigurationScenario):
                 N^param is the number of non-fixed parameters to be tuned.
                 Setting this parameter may make irace stop sooner than it should without
                 using all the available budget. We recommend to use the default value.
+            feature_data: FeatureDataFrame object with the feature data.
+                Currently not supported by IRACE.
         """
         """
         Other possible arguments that are not added yet to Sparkle:
@@ -324,6 +327,9 @@ class IRACEScenario(ConfigurationScenario):
             self.sparkle_objective = sparkle_objectives[0]
         else:
             self.sparkle_objective = None
+
+        if feature_data is not None:
+            print("WARNING: Instance features currently not supported by IRACE.")
 
         self.number_of_runs = number_of_runs
         self.solver_calls = solver_calls if solver_calls and solver_calls > 0 else None